+++ back/src/routers/certiRouter.js
@@ -2,6 +2,7 @@ import is from "@sindresorhus/is";
 import { Router } from "express";
 import { login_required } from "../middlewares/login_required";
 import { certiService } from "../services/certiService";
+import { certiMsg } from "../db/constant/errorMessage";
 
 const certiRouter = Router();
 
@@ -20,6 +21,15 @@ const createNewCertis = async (req, res, next) => {
     const { title, detail, date } = req.body;
     /* -req.currentUserId from login-requires -*/
     const user_id = req.currentUserId;
+
+    //ERROR THROW
+    if (!title) {
+      throw new Error(certiMsg.NO_TITLE_ERROR);
+    } else if (!detail) {
+      throw new Error(certiMsg.NO_DETAIL_ERROR);
+    } else if (!date) {
+      throw new Error(certiMsg.NO_DATE_ERROR);
+    }
 
     const newCerti = await certiService.createCertis({
       user_id,
@@ -44,6 +54,17 @@ const updateNewCerti = async (req, res, next) => {
     const obj_id = req.body._id;
     const { title, detail, date } = req.body;
 
+    //ERROR THROW
+    if (!title) {
+      throw new Error(certiMsg.NO_TITLE_ERROR);
+    } else if (!detail) {
+      throw new Error(certiMsg.NO_DETAIL_ERROR);
+    } else if (!date) {
+      throw new Error(certiMsg.NO_DATE_ERROR);
+    } else if (!obj_id) {
+      throw new Error(certiMsg.NO_OBJ_ERROR);
+    }
+
     //changed Input
     const newInput = { title, detail, date };
     const certi = await certiService.updateCerti(obj_id, newInput);
@@ -58,6 +79,11 @@ const updateNewCerti = async (req, res, next) => {
 const getCertis = async (req, res, next) => {
   try {
     const user_id = req.currentUserId;
+
+    if (!user_id) {
+      throw new Error(certiMsg.NO_USERID_ERROR);
+    }
+
     const certis = await certiService.getCertis(user_id);
     res.status(201).json(certis);
   } catch (err) {
@@ -69,6 +95,10 @@ const getCertis = async (req, res, next) => {
 const deleteCerti = async (req, res, next) => {
   try {
     const obj_id = req.body._id;
+
+    if (!obj_id) {
+      throw new Error(certiMsg.NO_OBJ_ERROR);
+    }
 
     //changed Input
     await certiService.deleteCerti(obj_id);
